+++ R/summaryByCrossValid.r
@@ -1,7 +1,7 @@
 #' Summarize distribution/niche model cross-validation object
 #'
 #' This function summarizes models calibrated using the \code{\link[enmSdm]{trainByCrossValid}} function. It returns aspects of the best models across k-folds (the particular aspects depends on the kind of models used).
-#' @param x An object of class \code{crossValid} (which is also a list). Note that the object \emph{must} include a sublist named \code{tuning}. For some models it must also include the sublist names \code{models}.
+#' @param x An object of class \code{crossValid} (which is also a list). Note that the object \emph{must} include a sublist named \code{tuning}.
 #' @param trainFxName Character, name of function used to train the SDM (examples: \code{'trainGlm'}, \code{'trainMaxEnt'}, \code{'trainBrt'})
 #' @param metric Metric by which to select the best model in each k-fold. This can be any of the columns that appear in the data frames in \code{x$tuning} (or any columns added manually), but typically is one of the following \emph{plus} either \code{Train}, \code{Test}, or \code{Delta} (e.g., \code{'logLossTrain'}, \code{'logLossTest'}, or \code{'logLossDelta'}):
 #' \itemize{
@@ -9,7 +9,6 @@
 #' 	\item \code{'cbi'}: Continuous Boyce Index (CBI). Calculated with \code{\link[enmSdm]{contBoyce}}.
 #' 	\item \code{'auc'}: Area under the receiver-operator characteristic curve (AUC). Calculated with \code{\link[enmSdm]{aucWeighted}}.
 #' 	\item \code{'tss'}: Maximum value of the True Skill Statistic. Calculated with \code{\link[enmSdm]{tssWeighted}}.
-#' 	\item \code{'fpb'}: Mean Fpb across thresholds from 0 to 1 in steps of 0.01. Calculated with \code{\link[enmSdm]{fpb}}. NOT USED AT PRESENT.
 #' 	\item \code{'msss'}: Sensitivity and specificity calculated at the threshold that maximizes sensitivity (true presence prediction rate) plus specificity (true absence prediction rate).
 #' 	\item \code{'mdss'}: Sensitivity (se) and specificity (sp) calculated at the threshold that minimizes the difference between sensitivity and specificity.
 #' 	\item \code{'orss'}: Maximum odds ratio skill score (Wunderlich et al. 2019).
@@ -17,12 +16,13 @@
 #' 	\item \code{'minTrainPres'}: Sensitivity and specificity at the greatest threshold at which all training presences are classified as "present".
 #' 	\item \code{'trainSe95'} and/or \code{'trainSe90'}: Sensitivity at the threshold that ensures either 95% or 90% of all training presences are classified as "present" (training sensitivity = 0.95 or 0.9).
 #' }
-#' @param descending Logical, if \code{TRUE} (default), for each k-fold sort models by the value listed in \code{metric} in descending order (highest connotes "best", lowest "worst"). If \code{FALSE} use the lowest value of \code{metric}.
+#' @param decreasing Logical, if \code{TRUE} (default), for each k-fold sort models by the value listed in \code{metric} in decreasing order (highest connotes "best", lowest "worst"). If \code{FALSE} use the lowest value of \code{metric}.
 #' @return Data frame with statistics on the best set of models across k-folds. Depending on the model algorithm, this could be:
 #' \itemize{
-#' 	\item GLMs: Frequency of use of each term in the best models
-#' 	\item Maxent: Mean regularization parameter and frequency of feature classes used in the best models (Maxent)
-#' 	\item BRTs: Learning rate, tree complexity, and bag fraction (BRTs).
+#' 	\item BRTs (boosted regression trees): Learning rate, tree complexity, and bag fraction.
+#' 	\item GLMs (generalized linear models): Frequency of use of each term in the best models.
+#' 	\item Maxent: Mean regularization parameter and frequency of feature classes used in the best models.
+#' 	\item NSs (natural splines): Data frame, one row per fold and one column per predictor, with values representing the maximum degrees of freedom used for each variable in the best model of each fold.
 #' }
 #' @seealso \code{\link[enmSdm]{trainByCrossValid}}
 #' @examples
@@ -81,11 +81,19 @@ summaryByCrossValid <- function(
 	x,
 	trainFxName = 'trainGlm',
 	metric = 'cbiTest',
-	descending = TRUE
+	decreasing = TRUE
 ) {
 
 	tuning <- x$tuning
 
+	### order each tuning attempt by performance
+	for (k in seq_along(tuning)) {
+	
+		tuningOrder <- order(tuning[[k]][ , metric], decreasing=decreasing)
+		tuning[[k]] <- tuning[[k]][tuningOrder, ]
+		
+	}
+	
 	### BRT
 	#######
 	if (trainFxName == 'trainBrt') {
@@ -95,9 +103,6 @@ summaryByCrossValid <- function(
 		for (k in seq_along(tuning)) {
 		
 			thisTuning <- tuning[[k]]
-			thisTuningOrder <- order(thisTuning[ , metric], decreasing=descending)
-			thisTuning <- thisTuning[thisTuningOrder, ]
-			
 			if (thisTuning$converged[1] & thisTuning$enoughTrees[1]) {
 				
 				params <- rbind(
@@ -140,8 +145,6 @@ summaryByCrossValid <- function(
 		for (k in seq_along(tuning)) {
 		
 			thisTuning <- tuning[[k]]
-			thisTuningOrder <- order(thisTuning[ , metric], decreasing=descending)
-			thisTuning <- thisTuning[thisTuningOrder, ]
 			
 			thisTerm <- thisTuning$model[1]
 			thisTerm <- strsplit(thisTerm, ' ')[[1]]
@@ -162,8 +165,6 @@ summaryByCrossValid <- function(
 		for (k in seq_along(tuning)) {
 		
 			thisTuning <- tuning[[k]]
-			thisTuningOrder <- order(thisTuning[ , metric], decreasing=descending)
-			thisTuning <- thisTuning[thisTuningOrder, ]
 			
 			thisTerm <- thisTuning$model[1]
 			thisTerm <- strsplit(thisTerm, ' ')[[1]]
@@ -182,7 +183,7 @@ summaryByCrossValid <- function(
 		out <- out[order(out$frequency, decreasing = TRUE), ]
 		out$proportionOfModels <- out$frequency / length(x$tuning)
 
-		rownames(out) <- 1:nrow(out)
+		rownames(out) <- paste0('fold', 1:nrow(out))
 
 	### MAXENT
 	##########
@@ -195,8 +196,6 @@ summaryByCrossValid <- function(
 		for (k in seq_along(tuning)) {
 		
 			thisTuning <- tuning[[k]]
-			thisTuningOrder <- order(thisTuning[ , metric], decreasing=descending)
-			thisTuning <- thisTuning[thisTuningOrder, ]
 			
 			thisTerm <- thisTuning[1 , c('linear', 'quadratic', 'product', 'hinge', 'threshold')]
 			feats <- rbind(feats, thisTerm)
@@ -227,9 +226,75 @@ summaryByCrossValid <- function(
 		out$proportionOfModels <- NA
 		featureRows <- out$term %in% c('linear', 'quadratic', 'hinge', 'product', 'threshold')
 		out$proportionOfModels[featureRows] <- out$value[featureRows] / length(x$tuning)
-		rownames(out) <- 1:nrow(out)
+		rownames(out) <- paste0('fold', 1:nrow(out))
 	
-	} # maxent
+	### NS
+	######
+	
+	} else if (trainFxName == 'trainNs') {
+		
+		# get predictor names
+		strings <- colnames(tuning[[1]])
+		strings <- strings[grepl(strings, pattern='splines::ns')]
+		strings <- strsplit(strings, 'splines::ns')
+		preds <- character()
+		for (i in seq_along(strings)) {
+			string <- strings[[i]][2]
+			pred <- substr(string, 2, regexpr(string, pattern=' df') - 2)
+			preds <- c(preds, pred)
+		}
+
+		preds <- sort(unique(preds))
+		
+		# create table to hold information on degrees of freedom for each top model
+		out <- data.frame(DUMMY=NA)
+		colnames(out) <- preds[1]
+		if (length(preds) > 1) {
+			for (i in 2:length(preds)) {
+				out$DUMMY <- NA
+				colnames(out)[ncol(out)] <- preds[i]
+			}
+		}
+		out <- out[rep(1, length(tuning)), ]
+		rownames(out) <- paste0('fold', 1:nrow(out))
+		
+		# find df for each predictor
+		for (k in seq_along(tuning)) {
+		
+			thisTuning <- tuning[[k]]
+
+			for (pred in preds) {
+			
+				colPattern <- paste0('splines::ns\\(', pred, ', df')
+				cols <- colnames(thisTuning)[grepl(colnames(thisTuning), pattern=colPattern)]
+
+				# get degrees of freedom used to evaluate this variable
+				dfs <- numeric()
+				for (countCol in seq_along(cols)) {
+					before <- paste0('splines::ns\\(', pred, ', df')
+					thisDf <- strsplit(cols[countCol], before)[[1]][2]
+					thisDf <- strsplit(thisDf, ' ')[[1]][3]
+					thisDf <- substr(thisDf, 1, nchar(thisDf) - 1)
+					thisDf <- as.numeric(thisDf)
+					dfs <- c(dfs, thisDf)
+				}
+				
+				vals <- thisTuning[1, cols, drop=FALSE]
+				bestDf <- c(!apply(vals, 1, FUN=is.na))
+				bestDf <- if (all(!bestDf)) {
+					NA
+				} else {
+					max(dfs[bestDf])
+				}
+				
+				out[k, pred] <- bestDf
+				
+			} # next predictor	
+				
+		} # next model
+		
+		
+	} # NSs
 	
 	out